Add negative RBAC label matches
Normal `node_labels` matches are positive only - you must provide an
exact label or a regexp that matches all valid labels.

This is problematic when you want to grant access to "any label value
except X". You could put "X" in the `deny` section of a role, but this
prevents any other roles from granting access to "X". For example, users
may have many dynamically-named environments and one "prod", and they
want to only grant access to "prod" via workflow API.

This commit adds support for `+regexp.not` function, for example
`+regexp.not(prod)`. The specific prefix was chosen because a valid
regexp can't start with `+`.